Rename mcp nacos implementation to nacos2 (#957)
