+++ auto-configurations/spring-ai-alibaba-autoconfigure-nacos2-mcp-client/src/main/resources/META-INF/spring-configuration-metadata.json
