+++ auto-configurations/spring-ai-alibaba-autoconfigure-nacos2-mcp-server/src/main/resources/META-INF/spring-configuration-metadata.json
